==> 3.Read_Results.ipynb <==
import os
import cPickle as pickle

resultsList = []
for item in os.listdir("permutations"):
    resultsList.append(pickle.load(open("permutations/" + item, "rb")))

numCorrect = 0
numTotal = 0
for result in resultsList:
    numCorrect += result['number_correct']
    numTotal += 2
print numCorrect / float(numTotal)
